+++ src/hal/PiezoToneOutput.h
@@ -15,5 +15,5 @@ private:
     int _pin;
     uint8_t _channel;
     uint8_t _resolutionBits;
-    uint32_t _toneHz = 2400;
+    uint32_t _toneHz = 3200;
 };